Remove SNR validation
Would prevent remediation with old templates
